css-transforms-1/transform_translate_zero_prefixed.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform supports translate(0, 0)"> 
    <script type="text/javascript" src="../../resources/testharness.js"></script>
    <script type="text/javascript" src="../../resources/testharnessreport.js"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
        // Set the transform 
        document.getElementById("test").style.webkitTransform = "translate(0, 0)";
        /* document.getElementById("test").style.mozkitTransform = "translate(0, 0)";
        document.getElementById("test").style.msTransform = "translate(0, 0)";
        document.getElementById("test").style.oTransform = "translate(0, 0)";
        document.getElementById("test").style.transform = "translate(0, 0)"; */
 
        // Verify that the transform was set as expected
        var value = document.getElementById("test").style.getPropertyCSSValue("-webkit-transform").cssText;
        /* var value = document.getElementById("test").style.getPropertyCSSValue("-moz-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("-ms-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("-0-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText; */
        
        test(function() {assert_equals(value,
                        "translate(0px, 0px)",
                        "transform should be translate(0px, 0px)")}, 
                        "transform_translate_0_0");
    </script>    
</body>
</html>
